# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: NinerLog API
  version: 1.0.0
  description: |
    RESTful API for EASA/FAA compliant pilot logbook management.
    
    ## Features
    - Multi-license support (PPL, SPL, CPL, ATPL, IR)
    - EASA and FAA regulatory compliance
    - Flight logging with detailed time tracking
    - Currency calculations
    - Statistics and hour totals per license
    
    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require JWT authentication.
    Include the token in the Authorization header: `Bearer <token>`
    
  contact:
    name: Frederic Jaeckel
    email: frederic@ninerlog.app
  license:
    name: TBD
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://app.ninerlog.com/api/v1
    description: Production server
...
paths:
...
  /auth/oidc/exchange:
    post:
      tags:
        - Authentication
      summary: Exchange an OIDC handoff code for tokens
      description: |
        Final step of the OIDC login. After the provider redirects the browser
        back to `/auth/oidc/callback`, the API sends it on to the frontend with
        a single-use `oidc_code` query parameter. The client posts that code
        here to receive the normal access/refresh token pair.

        The code is redeemable exactly once and expires within a minute, which
        is why the tokens themselves never appear in a URL.
      operationId: exchangeOidcCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  description: The single-use code from the `oidc_code` query parameter
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: The code is unknown, expired, or has already been redeemed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: The account has been disabled by an administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: OIDC is not configured on this server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /airports/{icaoCode}:
    get:
      tags: [Maps]
      summary: Get airport coordinates
      description: Get coordinates and metadata for an airport by ICAO code
      operationId: getAirport
      security:
        - bearerAuth: []
      parameters:
        - name: icaoCode
          in: path
          required: true
          description: ICAO airport code (4 characters)
          schema:
            type: string
            minLength: 4
            maxLength: 4
            example: EDDF
      responses:
        '200':
          description: Airport information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Airport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /airports/search:
    get:
      tags: [Maps]
      summary: Search airports
      description: Search airports by ICAO code prefix
      operationId: searchAirports
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query (ICAO code prefix)
          schema:
            type: string
            minLength: 1
            maxLength: 4
            example: EDD
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: List of matching airports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Airport'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reports/routes:
    get:
      tags: [Maps]
      summary: Get flight routes for map
      description: Get all unique flight routes with coordinates for map visualization
      operationId: getFlightRoutes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Flight routes with airport coordinates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightRoutesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reports/airport-stats:
    get:
      tags: [Maps]
      summary: Get airport statistics
      description: Get flight count and hours per airport for the authenticated user
      operationId: getAirportStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Airport statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AirportStats'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Import ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
...
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: pilot@example.com
        name:
          type: string
          example: John Doe
        createdAt:
          type: string
          format: date-time
          example: '2026-01-15T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-01-20T14:45:00Z'
        twoFactorEnabled:
          type: boolean
          example: false
          description: Whether 2FA is enabled for this account
        emailVerified:
          type: boolean
          example: true
          description: Whether the user has confirmed their email address.
        isAdmin:
          type: boolean
          example: false
          description: Whether this user is the platform admin (computed from ADMIN_EMAIL env var, not stored in DB)
        timeDisplayFormat:
          type: string
          enum: [hm, decimal]
          default: hm
          example: hm
          description: User's preferred time display format. "hm" for hours and minutes (1h 30m), "decimal" for decimal hours (1.5h).
        dateFormat:
          type: string
          enum: [DD.MM.YYYY, MM/DD/YYYY, YYYY-MM-DD]
          default: DD.MM.YYYY
          example: DD.MM.YYYY
          description: User's preferred date display format.
        decimalSeparator:
          type: string
          enum: [comma, dot]
          default: comma
          example: comma
          description: User's preferred decimal separator for number display. "comma" for European style (1,5), "dot" for US/UK style (1.5).
        preferredLocale:
          type: string
          enum: [en, de]
          default: en
          example: en
          description: User's preferred language for the interface.
        recencyPerModel:
          type: boolean
          default: true
          example: true
          description: >-
            Show informational 90-day landing recency (styled after EASA
            FCL.060(b)) per aircraft model/type.
        recencyPerRegistration:
          type: boolean
          default: false
          example: false
          description: >-
            Show informational 90-day landing recency per individual aircraft
            registration. Off by default — FCL.060(b) recency is defined per
            type/class, so per-registration tracking is purely a familiarity
            aid.
        flightListColumnMode:
          type: string
          enum: [auto, custom]
          default: auto
          example: auto
          description: >-
            How the optional columns of the flights list are chosen. "auto"
            lets the client pick them from the flights on the page, so an
            unused column (IFR time for a VFR-only pilot) never takes up width.
            "custom" means flightListColumns is the user's own choice.
        flightListColumns:
          type: array
          items:
            $ref: '#/components/schemas/FlightListColumn'
          default: []
          example: [picTime, nightTime, landings]
          description: >-
            The optional flights-list columns the user picked. Only meaningful
            when flightListColumnMode is "custom", where an empty array means
            "none of the optional columns". Stored deduplicated and in the
            canonical display order; unknown values are dropped.
    FlightListColumn:
      type: string
      description: >-
        An optional column of the flights list. Date, route, aircraft and total
        time are the identity of a logbook row and are always shown, so they
        are not part of this enum. The order below is the display order, and
        for the time columns also the priority order in which they survive as
        the list gets narrower.
      enum:
        - offOnBlock
        - picTime
        - nightTime
        - dualTime
        - ifrTime
        - crossCountryTime
        - sicTime
        - dualGivenTime
        - multiPilotTime
        - soloTime
        - simulatedFlightTime
        - function
        - landings
        - remarks
      example: picTime
...
    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          description: JWT access token (15 minutes validity)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT refresh token (7 days validity)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: Access token expiry in seconds
          example: 900
        user:
          $ref: '#/components/schemas/User'
...
    Airport:
      type: object
      required:
        - icao
        - name
        - latitude
        - longitude
      properties:
        icao:
          type: string
          description: ICAO airport code
          example: EDDF
        name:
          type: string
          description: Airport name
          example: Frankfurt am Main
        latitude:
          type: number
          format: double
          description: Latitude in decimal degrees
          example: 50.033333
        longitude:
          type: number
          format: double
          description: Longitude in decimal degrees
          example: 8.570556
        elevation:
          type: integer
          description: Airport elevation in feet
          example: 364
        country:
          type: string
          description: Country code (ISO 3166-1 alpha-2)
          example: DE
    FlightRoute:
      type: object
      required:
        - departureIcao
        - arrivalIcao
        - departureCoords
        - arrivalCoords
        - flightCount
      properties:
        departureIcao:
          type: string
          example: EDDF
        arrivalIcao:
          type: string
          example: EDDH
        departureCoords:
          type: object
          required: [lat, lng]
          properties:
            lat:
              type: number
              format: double
            lng:
              type: number
              format: double
        arrivalCoords:
          type: object
          required: [lat, lng]
          properties:
            lat:
              type: number
              format: double
            lng:
              type: number
              format: double
        flightCount:
          type: integer
          description: Number of flights on this route
          example: 5
    FlightRoutesResponse:
      type: object
      required:
        - routes
        - airports
      properties:
        routes:
          type: array
          items:
            $ref: '#/components/schemas/FlightRoute'
        airports:
          type: array
          items:
            $ref: '#/components/schemas/Airport'
    AirportStats:
      type: object
      required:
        - icao
        - name
        - latitude
        - longitude
        - departures
        - arrivals
        - totalFlights
      properties:
        icao:
          type: string
          example: EDDF
        name:
          type: string
          example: Frankfurt am Main
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        departures:
          type: integer
          description: Number of departures from this airport
        arrivals:
          type: integer
          description: Number of arrivals to this airport
        totalFlights:
          type: integer
          description: Total flights (departures + arrivals)

    # ── Crew & Contact schemas ───────────────────────────────────────────
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Validation failed
          description: Error message
        code:
          type: string
          example: email_not_verified
          description: |
            Optional machine-readable error code, used by clients to distinguish
            between conditions that share an HTTP status (e.g. `email_not_verified`
            vs `account_disabled` for 403 on /auth/login).
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: totalTime
              message:
                type: string
                example: Must be greater than 0
          description: Detailed validation errors
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            details:
              - field: totalTime
                message: Must be greater than 0
              - field: nightTime
                message: Cannot exceed total time
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
...